Fixing bug with multiple file comment
The problem occurred when you were using more then one file comment in your sass document.  The root cause was the fact that the FileDocumentation's raw_documentation method was depending on Documentation's raw_documentation (through super).  Documentation's raw_documentation method is designed to cache it's response therefore causing all file comments looked identical to the first one.  Instead of making it so that it does not cache itself, I changed FileDocumentation's raw_documentation logic to not depend on super.